+++ chain/chain/src/stateless_validation/mod.rs
@@ -1,4 +1,4 @@
 pub(crate) mod chunk_endorsement;
+pub mod chunk_validation;
 pub mod metrics;
 pub(crate) mod state_transition_data;
-pub mod state_witness;
